requirements.txt
altgraph==0.17.4
future==1.0.0
iso8601==2.1.0
macholib==1.16.3
packaging==24.0
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.4
pyserial==3.5
PyYAML==6.0.1
setuptools==69.5.1
tk==0.1.0